dev_requirements.txt
--index-url=https://packages.idmod.org/api/pypi/pypi-production/simple
livereload~=2.6.3
pytest-cache~=1.0
pytest-cov~=2.10.1
pytest-html~=3.1.1
pytest-runner~=5.2
pytest-timeout~=1.4.2
pytest-xdist~=2.2
pytest~=6.2.1
idmtools-models~=1.6.2
idmtools-test~=1.6.2
idmtools-platform-comps~=1.6.2